requirements.txt
Pillow==10.4.0
moviepy==1.0.3
